Initial refactor of directory structure
