+++ .idea/jsLinters/jslint.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project version="4">
+  <component name="JSLintConfiguration">
+    <option devel="true" />
+    <option es6="true" />
+    <option maxerr="50" />
+    <option node="true" />
+  </component>
+</project>
