+++ package.json
@@ -31,6 +31,8 @@
   },
   "homepage": "https://github.com/cassini-Inner/inner-souce-management#readme",
   "dependencies": {
+    "eslint": "^6.8.0",
+    "eslint-plugin-react": "^7.19.0",
     "react": "^16.13.0",
     "react-dom": "^16.13.0"
   },
