+++ postcss.config.js
@@ -1,7 +1,7 @@
-const tailwindcss = require('tailwindcss');
+const tailwindcss = require("tailwindcss");
 module.exports = {
     plugins: [
-        tailwindcss('./tailwind.config.js'),
-        require('autoprefixer')
+        tailwindcss("./tailwind.config.js"),
+        require("autoprefixer")
     ],
 };
