+++ src/Components/Common/InputFields/TextAreaInput.js
@@ -1,0 +1,12 @@
+import React from "react";
+
+export const TextAreaInput = (props) => {
+    return (
+        <textarea type="text" rows={props.rows ? props.rows : "5"}
+            cols={props.cols ? props.cols : "50"}
+            className={"pl-1 h-12 outline-none border-b-2 transition duration-300 focus:border-nebula-grey-600 placeholder-nebula-grey-500  " +
+                props.className}
+            placeholder={props.placeholder}
+            onKeyDown={props.onKeyDown ? props.onKeyDown : null}/>
+    );
+};
