+++ src/Components/Milestones/MilestoneCard.js
@@ -1,7 +1,8 @@
 import React, { Component } from 'react';
 
-import { StatusTag, InfoTag } from '../CommonComponents';
-import { EditIcon, ChevronDownIcon, DeleteIcon, ChevronUpIcon } from '../Icons';
+import { EditIcon, ChevronDownIcon, DeleteIcon, ChevronUpIcon } from '../Common/Icons';
+import { InfoTag } from '../Common/InfoTag/InfoTag'
+import StatusTags from '../Common/StatusTags/StatusTags'
 
 class MilestoneCard extends Component {
 
@@ -30,20 +31,18 @@ class MilestoneCard extends Component {
                     <svg className="w-3 relative mt-6" viewBox="0 0 12 12" fill="none" xmlns="http://www.w3.org/2000/svg">
                         <circle cx="6" cy="6" r="6" fill="#B3B3B3" />
                     </svg>
-                    <div className={(this.props.index === this.props.lastIndex - 1 ? "" : "h-full ") + " w-px mx-auto bg-nebula-grey-400"}></div>
+                    <div className={(this.props.index === this.props.lastIndex - 1 ? '' : 'h-full ') + ' w-px mx-auto bg-nebula-grey-400'}/>
                 </div>
                 <div className="mx-4 flex-1 ">
                     <div className="flex items-center flex-wrap">
                         <div className="flex flex-1 flex-col">
-
-                            <div className="h-4"></div>
+                            <div className="h-4"/>
                             <div className="flex-1">
                                 <p className="font-semibold text-nebula-grey-600">Milestone #{this.props.index + 1}</p>
                             </div>
-                            <div className="h-4"></div>
+                            <div className="h-4"/>
                         </div>
-                        {
-                            isEditMode &&
+                        {isEditMode &&
                             <div className="flex">
                                 <EditIcon className="text-nebula-blue mx-4" />
                                 <DeleteIcon className="text-nebula-red mx-4" />
@@ -63,7 +62,7 @@ class MilestoneCard extends Component {
                                 </button>
                             }
                         </div>
-                        <StatusTag statusTag={this.props.milestone.status}></StatusTag>
+                        <StatusTags statusTag={this.props.milestone.status}/>
                         {
                             isExpanded &&
                             <div className="" >
