+++ src/Components/routes/Home/Content.js
@@ -1,0 +1,16 @@
+import React, { Component } from "react";
+import JobList from "../../Jobs/JobList";
+import OngoingJobsGrid from "../../Jobs/OngoingJobsGrid";
+
+class Content extends Component {
+    render() {
+        return (
+            <div className="h-auto mt-4">
+                <OngoingJobsGrid maxCount = {1} location = "home" />
+                <JobList title = "Explore Jobs" setModalState = { this.props.setModalState } />
+            </div>
+        );
+    }
+}
+
+export default Content;
