+++ src/index.js
@@ -1,6 +1,6 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import App from './Components/App';
+import React from "react";
+import ReactDOM from "react-dom";
+import App from "./Components/App";
 import "../assets/style/index.css";
 
 ReactDOM.render(<App />, document.getElementById("App"));
